requirements.txt
annotated-types==0.6.0
blis==0.7.11
brotlipy==0.7.0
catalogue==2.0.10
certifi 
cffi 
charset-normalizer 
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==3.0.0
confection==0.1.3
contourpy==1.1.1
cryptography 
cycler==0.12.1
cymem==2.0.8
Farama-Notifications==0.0.4
fonttools==4.43.1
gym
gym-notices
gymnasium
hashids==1.3.1
jericho==3.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
langcodes==3.3.0
MarkupSafe==2.1.3
matplotlib==3.8.1
mementos==1.3.1
more-itertools==10.1.0
murmurhash==1.0.10
networkx==3.2.1
numpy
packaging==23.2
Pillow
pluggy
preshed==3.0.9
prompt-toolkit==3.0.39
pycosat 
pycparser 
pydantic==2.4.2
pydantic_core==2.10.1
pyOpenSSL 
pyparsing==3.1.1
PySocks 
python-dateutil==2.8.2
requests 
ruamel.yaml 
ruamel.yaml.clib 
six
smart-open==6.4.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
TatSu
thinc==8.2.1
toolz 
tqdm 
typer==0.9.0
typing_extensions==4.8.0
urllib3 
wasabi==1.1.2
wcwidth==0.2.9
weasel==0.3.3
zstandard 
hydra-core
gradio
torch
transformers==4.37.2
accelerate
peft
openai
wandb
beautifulsoup4
sentencepiece
pyinstrument
memory_profiler
appium-python-client
tenacity
termcolor
google-generativeai
jupyter
